+++ resources/sources/Input/FalconManager.cpp
@@ -239,7 +239,7 @@ float FalconManager::getNewPosition(int direction) {
 	// TODO direction checken ob 0 1 2 
 	// TODO 3 inch in z richtung
 	//printPosition();
-	return m_positionServo[direction]*100/(2*inch);
+	return float(m_positionServo[direction]*100/(2*inch));
 }
